refactor(logic): split mod.rs en traits.rs et tests.rs
- mod.rs 702 → 160 LOC
- extraction trait DatabaseInterface + LogicTrait → traits.rs (181 LOC)
- extraction #[cfg(test)] mod tests → tests.rs (365 LOC)
- zéro changement comportemental
